+++ apps/web/src/app/main/chat/Chat.tsx
@@ -4,9 +4,9 @@ import { motion } from 'motion/react';
 import { useEffect, useLayoutEffect, useState } from 'react';
 import type { z } from 'zod';
 import type { MessageMetadata } from '@/app/api/chat/messageMetadata';
-import { type ChannelSchema, DB } from '@/app/idb/db';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { useChat } from '@/core/react/useChat';
+import { type ChannelSchema, DB } from '@/idb/db';
 import { cn } from '@/lib/utils';
 import { generateUIMessage, messagesToThreads } from '../../../core/helper';
 import { currentChannelIdAtom } from '../store/currentChannelStore';
