+++ apps/web/src/app/main/header/EditChannelDialog.tsx
@@ -2,7 +2,6 @@ import { useSetAtom } from 'jotai';
 import { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 import type { z } from 'zod';
-import { type ChannelSchema, DB } from '@/app/idb/db';
 import { Button } from '@/components/ui/button';
 import {
   Dialog,
@@ -16,6 +15,7 @@ import {
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
+import { type ChannelSchema, DB } from '@/idb/db';
 import { channelListRefreshAtom } from '../store/channelListStore';
 
 interface EditChannelDialogProps {
