+++ apps/web/src/app/main/store/currentChannelStore.ts
@@ -1,4 +1,4 @@
 import { atom } from 'jotai';
-import { DEFAULT_CHANNEL_ID } from '@/app/idb/db';
+import { DEFAULT_CHANNEL_ID } from '@/idb/db';
 
 export const currentChannelIdAtom = atom(DEFAULT_CHANNEL_ID);
